Organize shared p2p code into a p2p package
